Client and Server work successfully. Here is how they work:
Request (5/10 pts)

Create a client program which allows the user to enter a text message to send to the server.  This message should be sent to the server program via socket.  Since this is a demo program, you can use the local 127.0.0.1 IP address for your socket.  You can designate the port

Links to an external site. to any open and valid number.
Response (5/10 pts)

After sending a request to the server, the client should listen for a response.  If it receives a response, print the received message text to the console.

Request (5/10 pts)

Create a server program which actively listens for requests from the client program.  If it receives a request, log it in the console.  Your server should use the same IP address and port number used for your client.
Response (5/10 pts)

After receiving a request from a client, the server should send a response.  The response text should be the uppercase version of the text it received.  Log the response in the console.